Fix bug in selecting listening port
Previously, the port was hard-coded to be "3000". This commit rectifies
that by selecting the port based on an environment variable (NODE_ENV). If
the environment is explicitly set to "development", we use the port "300".
Otherwise, production is assumed and we serve on port "80". The bug this
commit is referencing is located in the bugs section of commit ec46503.